Refactor languageName on page in favor of helper
